% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnaufel_rmd_format.R
\name{fnaufel_rmd_format}
\alias{fnaufel_rmd_format}
\title{Custom HTML template}
\usage{
fnaufel_rmd_format(toc = TRUE, ...)
}
\arguments{
\item{toc}{should a table of contents be displayed?}

\item{...}{additional arguments provided to \@code{html_document}}
}
\description{
Loads additional style and template file
}
